consolidate theme functionality in design
