+++ apps/tester-core/src/theme/CustomThemes.ts
@@ -3,7 +3,7 @@ import { Platform } from 'react-native';
 import { createAndroidTheme } from '@fluentui-react-native/android-theme';
 import { createAppleTheme } from '@fluentui-react-native/apple-theme';
 import { createDefaultTheme } from '@fluentui-react-native/default-theme';
-import { ThemeReference } from '@fluentui-react-native/theme';
+import { ThemeReference } from '@fluentui-react-native/design/theming';
 import type { ThemeOptions } from '@fluentui-react-native/design/theming';
 
 import type { OfficeBrand } from './applyBrand';
