+++ docs/pages/Guides/UpdateThemeProvider.md
@@ -22,7 +22,7 @@ registry.setTheme((parentTheme: Theme) => {
 the `ThemeReference` would have:
 
 ```ts
-import { ThemeReference } from '@fluentui-react-native/theme';
+import { ThemeReference } from '@fluentui-react-native/design/theming';
 import { Theme, PartialTheme } from '@fluentui-react-native/framework';
 
 const themeRef = new ThemeReference(
