+++ packages/agentic-design/package.json
@@ -32,6 +32,12 @@
       "import": "./lib/theming/index.js",
       "default": "./src/theming/index.ts"
     },
+    "./testing": {
+      "types": "./lib/testing/index.d.ts",
+      "react-native": "./src/testing/index.ts",
+      "import": "./lib/testing/index.js",
+      "default": "./src/testing/index.ts"
+    },
     "./package.json": "./package.json"
   },
   "scripts": {
@@ -43,6 +49,9 @@
     "test": "fluentui-scripts jest",
     "update-snapshots": "fluentui-scripts jest -u"
   },
+  "dependencies": {
+    "@fluentui-react-native/framework-base": "workspace:*"
+  },
   "devDependencies": {
     "@babel/core": "catalog:",
     "@fluentui-react-native/design-tokens-android": "catalog:",
@@ -51,6 +60,10 @@
     "@fluentui-react-native/design-tokens-win32": "catalog:",
     "@fluentui-react-native/design-tokens-windows": "catalog:",
     "@fluentui-react-native/scripts": "workspace:*",
+    "@react-native-community/cli": "^20.0.0",
+    "@react-native-community/cli-platform-android": "^20.0.0",
+    "@react-native-community/cli-platform-ios": "^20.0.0",
+    "@react-native/babel-preset": "^0.81.0",
     "@rnx-kit/tools-filesystem": "^0.2.0",
     "@types/react": "~19.1.4",
     "@types/react-test-renderer": "^19.1.0",
@@ -74,7 +87,7 @@
     }
   },
   "furn": {
-    "jestPlatform": "win32"
+    "jestPlatform": "macos"
   },
   "rnx-kit": {
     "kitType": "library",
