+++ packages/agentic-design/src/testing/index.ts
@@ -1,0 +1,1 @@
+export { mockTheme } from './mockTheme';
