+++ packages/agentic-design/src/theming/__tests__/themeReference.test.ts
@@ -1,7 +1,7 @@
-import { mockTheme } from '@fluentui-react-native/test-tools';
-import type { Theme, Spacing, PartialTheme } from '@fluentui-react-native/design/theming';
+import { mockTheme } from '../../testing';
+import type { Theme, Spacing, PartialTheme } from '../index';
 
-import { ThemeReference } from './themeReference';
+import { ThemeReference } from '../themeReference';
 
 const themeBase = mockTheme;
