+++ packages/agentic-design/src/theming/index.ts
@@ -30,3 +30,7 @@ export type {
 } from './Typography.types';
 export { ThemeContext, useTheme } from './context';
 export type { OfficePalette } from './palette.types';
+export { ThemeProvider } from './ThemeProvider';
+export type { ThemeProviderProps } from './ThemeProvider';
+export { ThemeReference } from './themeReference';
+export type { OnThemeChange, ThemeRecipe, ThemeTransform } from './themeReference';
