+++ packages/agentic-design/src/theming/mergeTheme.ts
@@ -1,5 +1,6 @@
 import { immutableMerge } from '@fluentui-react-native/framework-base';
-import type { Theme, PartialTheme } from '@fluentui-react-native/design/theming';
+
+import type { Theme, PartialTheme } from './index';
 
 export function mergeTheme<TTheme extends Theme, TPartial extends PartialTheme>(base: TTheme, partial: TPartial): TTheme {
   return immutableMerge(base, partial as unknown as TTheme);
