+++ packages/theming/android-theme/tsconfig.json
@@ -22,9 +22,6 @@
       "path": "../theme-tokens/tsconfig.json"
     },
     {
-      "path": "../../framework/theme/tsconfig.json"
-    },
-    {
       "path": "../theming-utils/tsconfig.json"
     }
   ]
