+++ packages/theming/apple-theme/package.json
@@ -39,7 +39,6 @@
     "@fluentui-react-native/design-tokens-macos": "catalog:",
     "@fluentui-react-native/experimental-appearance-additions": "workspace:*",
     "@fluentui-react-native/framework-base": "workspace:*",
-    "@fluentui-react-native/theme": "workspace:*",
     "@fluentui-react-native/theme-tokens": "workspace:*",
     "@fluentui-react-native/theming-utils": "workspace:*",
     "assert-never": "^1.2.1"
