+++ packages/theming/apple-theme/src/createAppleTheme.ios.ts
@@ -1,7 +1,7 @@
 import { Appearance, NativeEventEmitter } from 'react-native';
 
 import { NativeAppearanceAdditions } from '@fluentui-react-native/experimental-appearance-additions';
-import { ThemeReference } from '@fluentui-react-native/theme';
+import { ThemeReference } from '@fluentui-react-native/design/theming';
 import type { Theme } from '@fluentui-react-native/design/theming';
 
 import { getBaseAppleThemeIOS } from './appleTheme.ios';
