+++ packages/theming/apple-theme/src/createAppleTheme.macos.ts
@@ -1,6 +1,6 @@
 import { Appearance } from 'react-native';
 
-import { ThemeReference } from '@fluentui-react-native/theme';
+import { ThemeReference } from '@fluentui-react-native/design/theming';
 import type { Theme } from '@fluentui-react-native/design/theming';
 import { getCurrentAppearance } from '@fluentui-react-native/theming-utils';
 import { AccessibilityInfo } from 'react-native-macos';
