+++ packages/theming/apple-theme/src/createAppleTheme.ts
@@ -1,5 +1,5 @@
 import { createDefaultTheme } from '@fluentui-react-native/default-theme';
-import { ThemeReference } from '@fluentui-react-native/theme';
+import { ThemeReference } from '@fluentui-react-native/design/theming';
 
 export function createAppleTheme(): ThemeReference {
   console.warn('Platform is not supported by apple theme, using default theme');
