+++ packages/theming/default-theme/package.json
@@ -35,7 +35,6 @@
   "dependencies": {
     "@fluentui-react-native/design": "workspace:*",
     "@fluentui-react-native/framework-base": "workspace:*",
-    "@fluentui-react-native/theme": "workspace:*",
     "@fluentui-react-native/theme-tokens": "workspace:*",
     "@fluentui-react-native/theming-utils": "workspace:*",
     "assert-never": "^1.2.1"
