+++ packages/theming/default-theme/src/__tests__/defaultThemes.test.windows.ts
@@ -1,0 +1,13 @@
+import { defaultFluentTheme, defaultFluentDarkTheme, defaultFluentHighConstrastTheme } from '../defaultTheme';
+
+it('defaultFluentTheme', () => {
+  expect(defaultFluentTheme).toMatchSnapshot();
+});
+
+it('defaultFluentDarkTheme', () => {
+  expect(defaultFluentDarkTheme).toMatchSnapshot();
+});
+
+it('defaultFluentHighConstrastTheme', () => {
+  expect(defaultFluentHighConstrastTheme).toMatchSnapshot();
+});
