+++ packages/theming/default-theme/tsconfig.json
@@ -21,9 +21,6 @@
       "path": "../theme-tokens/tsconfig.json"
     },
     {
-      "path": "../../framework/theme/tsconfig.json"
-    },
-    {
       "path": "../theming-utils/tsconfig.json"
     }
   ]
