+++ packages/theming/win32-theme/package.json
@@ -37,7 +37,6 @@
     "@fluentui-react-native/design": "workspace:*",
     "@fluentui-react-native/design-tokens-win32": "catalog:",
     "@fluentui-react-native/framework-base": "workspace:*",
-    "@fluentui-react-native/theme": "workspace:*",
     "@fluentui-react-native/theme-tokens": "workspace:*",
     "@fluentui-react-native/theming-utils": "workspace:*"
   },
