+++ packages/theming/win32-theme/tsconfig.json
@@ -25,9 +25,6 @@
       "path": "../theme-tokens/tsconfig.json"
     },
     {
-      "path": "../../framework/theme/tsconfig.json"
-    },
-    {
       "path": "../theming-utils/tsconfig.json"
     }
   ]
