+++ packages/utils/test-tools/package.json
@@ -29,9 +29,11 @@
     "lint": "fluentui-scripts lint",
     "update-snapshots": "fluentui-scripts jest -u"
   },
+  "dependencies": {
+    "@fluentui-react-native/design": "workspace:*"
+  },
   "devDependencies": {
     "@babel/core": "catalog:",
-    "@fluentui-react-native/design": "workspace:*",
     "@fluentui-react-native/scripts": "workspace:*",
     "@react-native/babel-preset": "^0.81.0",
     "@types/jest": "^29.0.0",
